+++ src/helpers/Manifest.php
@@ -221,8 +221,9 @@ EOT;
             $prefix = self::$isHot
                 ? $config['devServer']['publicPath']
                 : $config['server']['publicPath'];
-            // If the module isn't a full URL, prefix it
-            if (!UrlHelper::isAbsoluteUrl($module)) {
+            $useAbsoluteUrl = $config['useAbsoluteUrl'];
+            // If the module isn't a full URL, prefix it as required
+            if ($useAbsoluteUrl && !UrlHelper::isAbsoluteUrl($module)) {
                 $module = self::combinePaths($prefix, $module);
             }
             // Resolve any aliases
@@ -230,8 +231,8 @@ EOT;
             if ($alias) {
                 $module = $alias;
             }
-            // Make sure it's a full URL
-            if (!UrlHelper::isAbsoluteUrl($module) && !is_file($module)) {
+            // Make sure it's a full URL, as required
+            if ($useAbsoluteUrl && !UrlHelper::isAbsoluteUrl($module) && !is_file($module)) {
                 try {
                     $module = UrlHelper::siteUrl($module);
                 } catch (Exception $e) {
@@ -505,6 +506,10 @@ EOT;
         $cacheDuration = Craft::$app->getConfig()->getGeneral()->devMode
             ? self::DEVMODE_CACHE_DURATION
             : null;
+        // If we're in `devMode` invalidate the cache immediately
+        if (Craft::$app->getConfig()->getGeneral()->devMode) {
+            self::invalidateCaches();
+        }
         // Get the result from the cache, or parse the file
         $cache = Craft::$app->getCache();
         $settings = Twigpack::$plugin->getSettings();
@@ -596,7 +601,11 @@ EOT;
         if ($devMode && !$soft) {
             throw new NotFoundHttpException($error);
         }
-        Craft::error($error, __METHOD__);
+        if (self::$isHot) {
+            Craft::warning($error, __METHOD__);
+        } else {
+            Craft::error($error, __METHOD__);
+        }
     }
 
     // Private Static Methods